Text.Regex.Posix

Description

Module that provides the Regex backend that wraps the C POSIX.2 regex api. This is the backend being used by the regex-compat package to replace Text.Regex.

The Text.Regex.Posix module provides a backend for regular expressions. If you import this along with other backends, then you should do so with qualified imports, perhaps renamed for convenience.

If the =~ and =~~ functions are too high level, you can use the compile, regexec, and execute functions from importing either Text.Regex.Posix.String or Text.Regex.Posix.ByteString. If you want to use a low-level CString interface to the library, then import Text.Regex.Posix.Wrap and use the wrap* functions.

This module is only efficient with ByteString only if it is null terminated, i.e. (Bytestring.last bs)==0. Otherwise the library must make a temporary copy of the ByteString and append the NUL byte.

A String will be converted into a CString for processing. Doing this repeatedly will be very inefficient.

Note that the posix library works with single byte characters, and does not understand Unicode. If you need Unicode support you will have to use a different backend.

When offsets are reported for subexpression captures, a subexpression that did not match anything (as opposed to matching an empty string) will have its offset set to the unusedRegOffset value, which is (-1).

Benchmarking shows the default regex library on many platforms is very inefficient. You might increase performace by an order of magnitude by obtaining libpcre and regex-pcre or libtre and regex-tre. If you do not need the captured substrings then you can also get great performance from regex-dfa. If you do need the capture substrings then you may be able to use regex-parsec to improve performance.

Documentation

getVersion_Text_Regex_Posix :: Version

module Text.Regex.Base

Wrap, for =~ and =~~, types and constants

data Regex

A compiled regular expression.

newtype ExecOption

A bitmapped CInt containing options for execution of compiled regular expressions. Option values (and their man 3 regexec names) are

execBlank which is a complete zero value for all the flags. This is the blankExecOpt value.

execNotBOL (REG_NOTBOL) can be set to prevent ^ from matching at the start of the input.

execNotEOL (REG_NOTEOL) can be set to prevent $ from matching at the end of the input (before the terminating NUL).

Constructors

ExecOption CInt

newtype CompOption

A bitmapped CInt containing options for compilation of regular expressions. Option values (and their man 3 regcomp names) are

compBlank which is a completely zero value for all the flags. This is also the blankCompOpt value.

compExtended (REG_EXTENDED) which can be set to use extended instead of basic regular expressions. This is set in the defaultCompOpt value.

compNewline (REG_NEWLINE) turns on newline sensitivity: The dot (.) and inverted set [^ ] never match newline, and ^ and $ anchors do match after and before newlines. This is set in the defaultCompOpt value.

compIgnoreCase (REG_ICASE) which can be set to match ignoring upper and lower distinctions.

compNoSub (REG_NOSUB) which turns off all information from matching except whether a match exists.

Constructors

CompOption CInt

compBlank :: CompOption

A completely zero value for all the flags. This is also the blankCompOpt value.

execBlank :: ExecOption

A completely zero value for all the flags. This is also the blankExecOpt value.

unusedRegOffset :: RegOffset

(=~) :: (RegexMaker Regex CompOption ExecOption source, RegexContext Regex source1 target) => source1 -> source -> target

(=~~) :: (RegexMaker Regex CompOption ExecOption source, RegexContext Regex source1 target, MonadFail m) => source1 -> source -> m target

execNotBOL :: ExecOption

execNotEOL :: ExecOption

compExtended :: CompOption

compIgnoreCase :: CompOption

compNoSub :: CompOption

compNewline :: CompOption